Refactor plugin interface
