+++ backend/src/cli/plugin_interface.py
@@ -1,30 +1,9 @@
-from abc import ABC, abstractmethod
+"""M\xf3dulo interno deprecado.
 
-class PluginInterface(ABC):
-    """Interfaz base que define el comportamiento de un plugin de la CLI.
+Este archivo se mantiene por compatibilidad pero las clases y funciones
+relacionadas con plugins se encuentran ahora en ``src.cli.plugin``.
+"""
 
-    Los plugins deben declarar nombre, versión, autor y una breve
-    descripción para que la CLI pueda mostrarlos correctamente.
-    """
+from .plugin import PluginInterface
 
-    #: Nombre del plugin o subcomando
-    name: str = ""
-
-    #: Versión del plugin
-    version: str = "0.1"
-
-    #: Autor del plugin
-    author: str = ""
-
-    #: Breve descripción del plugin
-    description: str = ""
-
-    @abstractmethod
-    def register_subparser(self, subparsers):
-        """Registra los argumentos del subcomando en el parser."""
-        raise NotImplementedError
-
-    @abstractmethod
-    def run(self, args):
-        """Ejecuta la lógica del plugin."""
-        raise NotImplementedError
+__all__ = ["PluginInterface"]
