+++ backend/src/cli/plugin_loader.py
@@ -1,67 +1,13 @@
-import logging
-from importlib import import_module
-from importlib.metadata import entry_points
+"""M\xf3dulo interno deprecado.
 
-from .commands.base import BaseCommand
-from .plugin_interface import PluginInterface
-from .plugin_registry import registrar_plugin
+Las utilidades de carga y las clases para plugins ahora residen en
+``src.cli.plugin``.
+"""
 
+from .plugin import PluginCommand, descubrir_plugins, cargar_plugin_seguro
 
-class PluginCommand(BaseCommand, PluginInterface):
-    """Clase base para implementar comandos externos mediante plugins."""
-
-
-def descubrir_plugins():
-    """Descubre e instancia los plugins registrados bajo ``cobra.plugins``."""
-    plugins = []
-    try:
-        eps = entry_points(group="cobra.plugins")
-    except TypeError:  # Compatibilidad con versiones antiguas
-        eps = entry_points().get("cobra.plugins", [])
-
-    for ep in eps:
-        instancia = cargar_plugin_seguro(ep.value)
-        if instancia is not None:
-            plugins.append(instancia)
-    return plugins
-
-
-def cargar_plugin_seguro(ruta: str):
-    """Carga de forma segura un plugin a partir de ``modulo:Clase``."""
-    try:
-        module_name, class_name = ruta.split(":", 1)
-    except ValueError:
-        logging.error(f"Ruta de plugin inválida: {ruta}")
-        return None
-
-    try:
-        module = import_module(module_name)
-    except Exception as exc:
-        logging.error(f"Error importando módulo {module_name}: {exc}")
-        return None
-
-    try:
-        plugin_cls = getattr(module, class_name)
-    except AttributeError:
-        logging.error(f"No se encontró la clase {class_name} en {module_name}")
-        return None
-
-    if not issubclass(plugin_cls, PluginInterface):
-        logging.warning(
-            f"El plugin {ruta} no implementa PluginInterface"
-        )
-        return None
-
-    try:
-        instancia = plugin_cls()
-    except Exception as exc:
-        logging.error(f"Error instanciando plugin {ruta}: {exc}")
-        return None
-
-    if not getattr(instancia, "name", ""):
-        logging.warning(f"Plugin {ruta} no define atributo name")
-        return None
-
-    version = getattr(instancia, "version", "0")
-    registrar_plugin(instancia.name, version)
-    return instancia
+__all__ = [
+    "PluginCommand",
+    "descubrir_plugins",
+    "cargar_plugin_seguro",
+]
