+++ backend/src/tests/test_cli_plugins_cmd.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 import importlib.metadata
 
 from src.cli.cli import main
-from src.cli.plugin_loader import PluginCommand
+from src.cli.plugin import PluginCommand
 
 
 class DummyPlugin(PluginCommand):
@@ -35,7 +35,7 @@ def test_cli_plugins_muestra_registro():
         value="src.tests.test_cli_plugins_cmd:DummyPlugin",
         group="cobra.plugins",
     )
-    with patch("src.cli.plugin_loader.entry_points", return_value=importlib.metadata.EntryPoints((ep,))):
+    with patch("src.cli.plugin.entry_points", return_value=importlib.metadata.EntryPoints((ep,))):
         with patch("sys.stdout", new_callable=StringIO) as out:
             main(["plugins"])
     assert "dummy 5.2" in out.getvalue().strip()
@@ -47,7 +47,7 @@ def test_cli_plugin_ejemplo_hola():
         value="src.tests.test_cli_plugins_cmd:HolaPlugin",
         group="cobra.plugins",
     )
-    with patch("src.cli.plugin_loader.entry_points", return_value=importlib.metadata.EntryPoints((ep,))):
+    with patch("src.cli.plugin.entry_points", return_value=importlib.metadata.EntryPoints((ep,))):
         with patch("sys.stdout", new_callable=StringIO) as out:
             main(["hola"])
     assert "¡Hola desde un plugin!" in out.getvalue().strip()
