+++ examples/plugins/saludo_plugin.py
@@ -1,4 +1,4 @@
-from backend.src.cli.plugin_loader import PluginCommand
+from src.cli.plugin import PluginCommand
 
 class SaludoCommand(PluginCommand):
     """Comando de ejemplo que imprime un saludo."""
